third_party/blink/web_tests/platform/generic/fast/forms/time-multiple-fields/time-multiple-fields-narrow-width-scroll-expected.txt

Check if changing focus in a multiple fields input scrolls the content.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS initialScrollLeft < container.scrollLeft is true
PASS successfullyParsed is true

TEST COMPLETE
